refactor(FE): ReviewScroll에 맞추어 PostBar도 Scroll로 변경 (#146)
