+++ client/src/components/PostScroll/Post/PostFooter/ReviewButton/ReviewButton.tsx
@@ -2,7 +2,7 @@ import React, { useCallback, useContext } from "react";
 import ForumIcon from "@mui/icons-material/Forum";
 import useCommonModalStore from "@/store/useCommonModalStore";
 import ReviewModal from "@/components/Modal/ReviewModal/ReviewModal";
-import { PostContext } from "@/components/PostBar/Post/Post";
+import { PostContext } from "@/components/PostScroll/Post/Post";
 
 const ReviewButton = (): JSX.Element => {
   const { id: postId, code, language } = useContext(PostContext);
