+++ client/src/components/PostScroll/Post/PostTitle/AuthorProfile/FollowButton/FollowButton.scss
@@ -2,23 +2,23 @@
 
 .follow-button {
   @include green-button;
+  width: fit-content;
+  height: 1.4rem;
 
   font-size: 0.8rem;
-  height: 1.4rem;
-  width: fit-content;
 
   &--is-following {
     @include weview-button;
-
-    font-size: 0.75rem;
-    height: 1.5rem;
     width: fit-content;
+    height: 1.5rem;
 
     color: $line-color !important;
+
+    font-size: 0.75rem;
     border: 1px solid $line-color;
     &:hover {
-      background-color: $line-color;
       color: $weview-off-white !important;
+      background-color: $line-color;
     }
   }
 }
