+++ client/src/components/PostScroll/PostScroll.scss
@@ -1,21 +1,20 @@
 @import "@/constants/_library.scss";
 @import "@/styles/_global-styles.scss";
 
-.post-bar {
-  flex: none;
-
+.post-scroll {
   display: flex;
+  flex: none;
   flex-direction: column;
-  align-items: center;
-
-  margin: 0 $post-bar-margin-side;
   gap: $post-gap;
-
-  background-color: $weview-off-white;
-  overflow-y: scroll;
+  align-items: center;
 
   width: $main-post-bar-width;
   height: 100%;
+
+  margin: 0 $post-bar-margin-side;
+  overflow-y: scroll;
+
+  background-color: $weview-off-white;
 
   &::-webkit-scrollbar {
     display: none;
